// apps/frontend/src/lib/api/generated/models/OptionPrice.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Serializer for option price
 */
export type OptionPrice = {
    readonly id: number;
    readonly city_slug: string;
    readonly city_title: string;
    readonly technic_category_id: number | null;
    readonly technic_category_title: string | null;
    amount: string;
};